+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## upcoming
 
+- Adds `Checkbox`, `CheckboxBlank`, `CheckboxBlankDark` to `@graph-paper/icons` ([#64](https://github.com/graph-paper-org/graph-paper/pull/64))
 - `Button` component now accept `href` ([#59](https://github.com/graph-paper-org/graph-paper/pull/59))
 - adds `@graph-paper/portal` ([#55]((https://github.com/graph-paper-org/graph-paper/pull/55/))
 - adds `@graph-paper/optionmenu` ([#19](https://github.com/graph-paper-org/graph-paper/pull/19/))